further cleanups, dedupes from pairing
