+++ packages/kuma-gui/src/app/fake-fs/index.ts
@@ -3,7 +3,7 @@ import { mswHandlers } from '@kumahq/fake-api/msw'
 import type { ServiceDefinition, Token } from '@/services/utils'
 import { token } from '@/services/utils'
 import type { EndpointDependencies, FS } from '@/test-support'
-import { dependencies } from '@/test-support/fake'
+import { dependencies } from '@/test-support'
 
 const $ = {
   fakeFS: token<FS>('fake.fs'),
