+++ packages/kuma-gui/src/test-support/data/index.ts
@@ -1,7 +1,7 @@
 import { server as fakeApi } from '@kumahq/fake-api/msw'
 import { test as _test } from 'vitest'
 
-import { dependencies } from '../fake'
+import { dependencies } from '../'
 
 type FakeApiParameters = Parameters<typeof fakeApi<typeof dependencies>>
 export const server = (mock: FakeApiParameters[0], options: FakeApiParameters[1]) => fakeApi(mock, options, dependencies)
