+++ packages/kuma-gui/vite.config.production.ts
@@ -8,7 +8,7 @@ import { defineConfig } from 'vite'
 import svgLoader from 'vite-svg-loader'
 
 import { hoistUseStatements } from './dev-utilities/hoistUseStatements'
-import { dependencies } from './src/test-support/fake'
+import { dependencies } from './src/test-support'
 import { fs } from './src/test-support/mocks/fs'
 import { replicateKumaServer } from './vite.plugins'
 import type { UserConfigFn } from 'vite'
